+++ README.md
@@ -1,0 +1,32 @@
+
+# 2022级机器人
+
+
+## 贡献指南
+在进行更改之前，请阅读 [贡献指南](https://github.com/OpenHUTB/bazaar/blob/master/CONTRIBUTING.md) 文档。
+
+## 参与人
+| 学号         | 姓名  | 课设1 | 课设2 |
+|------------|-----|-----| -----|
+| 2209040021 | 周恒 |  |  |
+| 2209040022 | 罗亮 |编写一万年年历系统  |24点游戏设计  |
+| 2209040023 | 周若弈 |  |  |
+| 2209040024 | 马思凡 |  |  |
+| 2209040025 | 马嘉骏 |  |  |
+| 2209040026 | 唐晓南 |  |  |
+| 2209040027 | 刘文杰 |学生通讯录管理系统设计  |学生籍贯信息记录簿设计  |
+| 2209040028 | 宋俊涛 |  |  |
+| 2209040029 | 刘炫祺 |歌手比赛系统设计  | 学校运动会管理系统设计 |
+| 2209040030 | 伍常弘 |  |  |
+| 2209040031 | 丁清亮 |  |  |
+| 2209040032 | 陈炬 |销售管理系统设计  | 24点游戏设计 |
+| 2209040033 | 张寅 | 人机猜数游戏设计 | 小学生测验 |
+| 2209040034 | 姚昆利 |小学生测验|学生成绩管理系统设计|
+| 2209040035 | 吕银涛 |24点游戏设计  | 小学生测验|
+| 2209040036 | 江涛 |学校运动会管理系统设计  |学生信息管理系统设计  |
+| 2209040037 | 黄嘉宇 |学生信息管理系统设计  | 二十四点游戏设计 |
+| 2209040038 | 戴熙雯 | 学生成绩管理系统设计 | 实验设备管理系统设计 |
+| 2209040039 | 宁志 | 超市管理系统设计 | 学生信息管理系统设计|
+| 2209040040 | 高安南 | 实验设备管理系统设计 | 二十四点游戏设计 |
+
+
